refactor(sync-engine): hollow out the facade — behavior moves to its owning modules
create-rxdb-sync-engine.ts drops from 2,517 to 1,927 lines. The 17-member
public interface is unchanged; the ~20-file end-to-end facade test suite
is untouched and green — it is the oracle for these three moves:

- write-path/conflict-resolution.ts: the 289-line resolveConflict
  implementation leaves the facade's object literal; conflicts()/
  resolveConflict() stay byte-identical as thin delegations (the
  keep-ruling on that interface is respected — implementation only)
- local-coverage/reconcile-port.ts: openScopeDatabase's inline
  existence-reconcile port (dirtyWooIds / removeTargeted /
  pullTargetedAndPopulateManifest) becomes a factory with explicit deps
- maintenance/lane-registry.ts: ONE table now carries lane identity,
  cadence, dispatch target, ownership, static collection attribution,
  and all three orderings (seed retick, rebaseline retick, timer arming);
  the facade and maintenance-lanes both derive from it — replacing 4
  hand-correlated declarations across 13 edit sites in 2 naming
  conventions

Also declares the optional 'rebaselined' field on SyncReport honestly —
the facade's dispatch read it through an accidental type narrowing.

New glue: 141 non-test lines; everything else is relocation.
